continued greeks by category code changes

+++ RequestForQuoteInterfacesLibrary/ServiceInterfaces/IReportDataManager.cs
@@ -6,6 +6,6 @@ namespace RequestForQuoteInterfacesLibrary.ServiceInterfaces
     public interface IReportDataManager
     {
         void CompileRequestCountPerCategoryReport(string reportType, string categoryType, DateTime fromDate, int minimumCount);
-        void CompileGreeksPerCategoryReport(string reportType, string categoryType, ISet<string> greeks, DateTime maturityDateFrom, DateTime maturityDateTo, decimal minimumGreek);
+        void CompileGreeksPerCategoryReport(string reportType, string categoryType, ISet<string> greeks, DateTime maturityDateFrom, DateTime maturityDateTo, double minimumGreek);
     }
 }
